+++ pmc_automation_tools/__init__.py
@@ -12,7 +12,7 @@ from pmc_automation_tools.driver.common import (
     EXISTS
 )
 
-__version__ = "0.2.2"
+__version__ = "0.3.0"
 __all__ = [
     "UXDataSource",
     "UXDataSourceInput",